==> notebooks/03_feature_extraction_tfidf.ipynb <==
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import LabelEncoder
import joblib

df = pd.read_csv('../data/processed/sentiment140_clean.csv')
texts = df['clean_text'].astype(str).values
labels = df['sentiment'].values

label_encoder = LabelEncoder()
y = label_encoder.fit_transform(labels)
joblib.dump(label_encoder, '../models/label_encoder.pkl')


vectorizer = TfidfVectorizer(
    max_features=10000,
    ngram_range=(1, 2),
    sublinear_tf=True,
    norm='l2'
)
X = vectorizer.fit_transform(texts)
joblib.dump(vectorizer, '../models/tfidf_vectorizer.pkl')


joblib.dump((X, y), '../models/tfidf_features_labels.pkl')
