==> notebooks/04_model_training_logreg_svm.ipynb <==
import joblib
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.svm import LinearSVC
from sklearn.metrics import accuracy_score, f1_score, confusion_matrix


X, y = joblib.load('../models/tfidf_features_labels.pkl')

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, stratify=y, random_state=42
)


logreg = LogisticRegression(max_iter=1000, C=1.0)
logreg.fit(X_train, y_train)
y_pred_logreg = logreg.predict(X_test)
joblib.dump(logreg, '../models/logreg_model.pkl')


svm = LinearSVC(C=1.0)
svm.fit(X_train, y_train)
y_pred_svm = svm.predict(X_test)
joblib.dump(svm, '../models/svm_model.pkl')


acc_svm = accuracy_score(y_test, y_pred_svm)
f1_svm = f1_score(y_test, y_pred_svm)
conf_svm = confusion_matrix(y_test, y_pred_svm)
(acc_svm, f1_svm, conf_svm)
