==> notebooks/08_hyperparameter_tuning.ipynb <==
import joblib
import numpy as np
from sklearn.svm import LinearSVC
from sklearn.model_selection import GridSearchCV, StratifiedKFold


data = joblib.load("../models/tfidf_features_labels.pkl")
X, y = data


cv = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)


params = {
    "C": [0.01, 0.1, 1, 10, 100]
}


grid = GridSearchCV(
    estimator=LinearSVC(),
    param_grid=params,
    cv=cv,
    scoring="accuracy",
    n_jobs=-1
)


grid.fit(X, y)


print(grid.best_params_)
print(grid.best_score_)


joblib.dump(grid.best_estimator_, "../models/linear_svc_model.pkl")
